Switch media fw permissions checks to AttributionSource (wilhelm)
Attribution source is the abstraction to capture the data
flows for private data across apps. Checking permissions
for an attribution source does this for all apps in the
chain that would receive the data as well as the relevant
app ops are checked/noted/started as needed.

bug: 158792096

Test: atest CtsMediaTestCases
      atest CtsPermissionTestCases
      atest CtsPermission2TestCases
      atest CtsPermission3TestCases
      atest CtsPermission4TestCases
      atest CtsPermission5TestCases

Merged-In: I2061d398e8a96fb707ddf42c0412db6e7c68fe7f

Change-Id: I2061d398e8a96fb707ddf42c0412db6e7c68fe7f