Take local copy of CallbackProtector in Callbacks
We clear the sp holding the callback protector when destroying the
player object. Copy an sp of the protector into the callback objects,
which may still be firing as destruction is requested. Note, the
pre-destroy method tells the protector to drop all future callbacks, so
upstream of the callback class, no callbacks are received. Note,
libaudioclient only holds wp<> to the callback classes.

The CallbackProtector associated with a particular player does not
change over the lifetime of the player.

Test: Compiles
Bug: 361799177
Flag: EXEMPT bugfix
Change-Id: Icdd6179e04d339425a2ecf586d198ed236edd7c8